remove MapObject::Kind
That enum was an artefact of how things are structured in MainWindow,
and the map model/controller objects should really be oblivious of
that. Instead we now have MapObject::Group, which corresponds to the
object groups the game actually defines, and some custom handling of
unit types in Mainwindow and ObjectEditWidget.